+++ park-project/src/main/java/com/kielbiowski/parkproject/service/model/RequestService.java
@@ -1,10 +1,11 @@
-package com.kielbiowski.parkproject.service;
+package com.kielbiowski.parkproject.service.model;
 
 import com.kielbiowski.parkproject.dto.CarDTO;
 import com.kielbiowski.parkproject.dto.RequestDTO;
 import com.kielbiowski.parkproject.exception.NotFoundException;
 import com.kielbiowski.parkproject.model.Request;
 import com.kielbiowski.parkproject.repository.RequestRepository;
+import com.kielbiowski.parkproject.service.security.ServiceInterface;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
